requirements_dev.txt
numpy==1.24.2
matplotlib==3.8.0
pytest==7.4.2
tox==4.11.3
scipy==1.11.3
scikit-learn==1.3.1
libpysal==4.8.0
d-geolib==1.3.3
topojson==1.5
shapely==2.0.1
git+https://github.com/DataFusionTools/core
git+https://github.com/DataFusionTools/spatial_utils
git+https://github.com/DataFusionTools/interpolation